requirements.txt
oteapi-core==0.0.4
